Fix issue #1: Missing type check in jtfdep
This bug is problematic because of the convention that unassigned names are
assumed to be verbs, which is assumed to be the case in many places. However
when combined with gerunds, later assignment to the name of a non-verb type will
cause problems if not explicity checked. The source structure leaves no obvious
place for this check. If checked during gerund formation then the
unassigned name == verb "feature" won't work. Also, there are places outside of
the primitives where gerunds are "inspected" before use, so trying to put the
check in every primitive, in addition to being tedious and error prone, might
not cover all cases. This fix simply puts the check at the site of a crash
casued by certain random bit patterns in jtfdep. Other missing checks may still
be at large.

Example of code that can cause a crash

verb=: 0:`NOUN@.0
NOUN=: 'string',LF
verb''