Add experimental CSV file parser foreign
This is a proof-of-concept CSV parser. The interface, behavior and
performance characteristics will certainly change as development
continues.

X 16!:1 Y - Reads the csv file given by the file name Y and parses it
according to the options given in X. At the moment X is a literal list
of column types, which can be 'l' for literal and 'n' for numeric. If
an error occurs during numeric conversion a literal is returned
instead. For example:

'lnn' 16!:1 <'data.csv'

reads the file data.csv which is assumed to have 3 columns and returns
a boxed J table with the first column as literals and the remaining
columns as numeric (or literal in the case an invalid number was
specified.)

Future developments will include a writer, parsing into an inverted
table and more control over the parsing parameters.